fix(test): share SDK query mock across all test files
Extract the SDK `mock.module('@anthropic-ai/claude-agent-sdk')` registration
into `_shared-sdk-query-mock.ts` so both `sdk-integration.test.ts` and the
executor tests (`_sdk-mocks.ts`) use the same `queryMock` instance.

Previously, each file registered its own competing mock. Since bun's
`mock.module()` is process-global and first-registration-wins, whichever
file loaded first locked the global cache — making the other file's mock
assertions fail (0 calls). This caused 11 persistent test failures in the
full suite that passed individually.
